move getArgsFromRenameFunction to esql package
