+++ src/platform/packages/shared/kbn-esql-utils/index.ts
@@ -43,6 +43,7 @@ export {
   replaceESQLQueryIndexPattern,
   getCategorizeColumns,
   extractCategorizeTokens,
+  getArgsFromRenameFunction,
 } from './src';
 
 export { ENABLE_ESQL, FEEDBACK_LINK } from './constants';
