+++ src/platform/packages/shared/kbn-esql-utils/src/index.ts
@@ -25,6 +25,7 @@ export {
   getESQLQueryVariables,
   fixESQLQueryWithVariables,
   getCategorizeColumns,
+  getArgsFromRenameFunction,
 } from './utils/query_parsing_helpers';
 export { queryCannotBeSampled } from './utils/query_cannot_be_sampled';
 export {
